ibsim: fix size of slid/dlid fields in sim_request structure
The code treats them as 16-bit fields (by using htons), make
the structure match.  This also avoids the 32 bits of padding
before the length field.

+++ include/ibsim.h
@@ -62,8 +62,8 @@ struct sim_port {
 #define SIM_CTL_MAX_DATA	64
 
 struct sim_request {
-	uint32_t dlid;
-	uint32_t slid;
+	uint16_t dlid;
+	uint16_t slid;
 	uint32_t dqp;
 	uint32_t sqp;
 	uint32_t status;
